change the modal position and remove unused code
