+++ js/photo_booth.js
@@ -1,11 +1,8 @@
-import { appendCh, createEl } from "./utilities/dom.js";
+import { appendCh, createEl } from "./dom-utils.js";
 
 let isCameraWorking = true;
 
 export const createPhotoBoothEl = () => {
-  console.log("photo booth.... ");
-
-  /** create photo booth window */
   const video = createEl("video");
   video.setAttribute("autoplay", "");
 
@@ -20,9 +17,23 @@ export const createPhotoBoothEl = () => {
   appendCh(videoIcon, document.querySelector(".photo--menu__btn"));
 };
 
+const stopCamera = async (stream) => {
+  console.log("stopCamera...");
+
+  try {
+    if (stream) {
+      const tracks = await stream.getTracks();
+      tracks.forEach((track) => track.stop());
+      isCameraWorking = false;
+    }
+  } catch (err) {
+    console.log("Something went wrong...", err);
+  }
+};
+
 export const runPhotoBooth = async () => {
   const video = document.querySelector("video");
-  /** run webcam */
+
   let stream;
   try {
     stream = await navigator.mediaDevices.getUserMedia({ video: true });
@@ -31,84 +42,15 @@ export const runPhotoBooth = async () => {
     console.log("Error accessing camera: ", err);
   }
 
-  /** toggle start / stop buttons */
-  const toggleBtn = document.querySelector(".photo--menu__btn");
   const closeModalBtn = document.querySelector("#closeBtn");
   closeModalBtn.addEventListener("click", () => {
-    console.log("close button clicked");
-    // stopCamera(stream);
+    if (isCameraWorking) stopCamera(stream);
     isCameraWorking = !isCameraWorking;
-
-    if (stream) {
-      const tracks = stream.getTracks();
-      tracks.forEach((track) => track.stop());
-      video.srcObject = null;
-      isCameraWorking = false;
-    }
-    toggleVideoButtonClass();
   });
 
-  toggleBtn.addEventListener("click", () => {
-    console.log("toggleBtn button clicked");
-    // stopCamera(stream);
-    // if (stream) {
-    //   const tracks = stream.getTracks();
-    //   tracks.forEach((track) => track.stop());
-    //   videoElement.srcObject = null;
-
-    //   isCameraWorking = false;
-    //   toggleVideoButtonClass();
-    // }
-
+  const stopBtn = document.querySelector(".photo--menu__btn");
+  stopBtn.addEventListener("click", () => {
+    if (isCameraWorking) stopCamera(stream);
     isCameraWorking = !isCameraWorking;
-
-    if (stream) {
-      const tracks = stream.getTracks();
-      tracks.forEach((track) => track.stop());
-      video.srcObject = null;
-      isCameraWorking = false;
-    }
-    toggleVideoButtonClass();
   });
-};
-
-const stopCamera = (stream) => {
-  const video = document.querySelector("video");
-  console.log(stream, "--- stream? stopCamera");
-  isCameraWorking = !isCameraWorking;
-
-  if (stream) {
-    const tracks = stream.getTracks();
-    tracks.forEach((track) => track.stop());
-
-    video.srcObject = null;
-
-    isCameraWorking = false;
-  }
-  toggleVideoButtonClass();
-};
-
-const toggleVideoButtonClass = () => {
-  console.log(isCameraWorking, "-- toggle button triggered..");
-
-  const videoIcon = document.querySelector("#videoIcon");
-
-  if (isCameraWorking) {
-    videoIcon.classList.contains("photo--menu__btn__start") &&
-      videoIcon.classList.remove("photo--menu__btn__start");
-    videoIcon.classList.add("photo--menu__btn__stop");
-  } else {
-    videoIcon.classList.contains("photo--menu__btn__stop") &&
-      videoIcon.classList.remove("photo--menu__btn__stop");
-    videoIcon.classList.add("photo--menu__btn__start");
-  }
-};
-
-const playCamera = async (stream) => {
-  try {
-    stream = await navigator.mediaDevices.getUserMedia({ video: true });
-    video.srcObject = stream;
-  } catch (err) {
-    console.err("Error accessing camera: ", err);
-  }
 };
