+++ style.css
@@ -124,23 +124,17 @@ header .utility {
 header .menu__item,
 header .utility__item {
   display: flex;
-  justify-content: space-between;
-  padding-right: 0.7rem;
-  margin: 0 auto;
+  justify-content: center;
+  margin-right: 0.7rem;
 }
 header .menu__item i,
 header .utility__item i {
-  transform: scale(1.3);
-  width: 0.8rem;
-  height: auto;
+  transform: scale(1.1);
 }
-header .menu__item:nth-of-type(1),
-header .utility__item:nth-of-type(1) {
+header .menu__item:nth-of-type(1) {
   padding: 0.3rem 0.8rem;
 }
-header .menu__item:nth-of-type(1):hover, header .menu__item:nth-of-type(1):active,
-header .utility__item:nth-of-type(1):hover,
-header .utility__item:nth-of-type(1):active {
+header .menu__item:nth-of-type(1):hover, header .menu__item:nth-of-type(1):active {
   background-color: rgba(255, 255, 255, 0.2);
   border-radius: 4px;
 }
